Fix: Simplify commands cut copy delete (#11612)

+++ src/Files.App/Actions/FileSystem/DeleteItemAction.cs
@@ -4,10 +4,7 @@ using Files.App.Commands;
 using Files.App.Contexts;
 using Files.App.Extensions;
 using Files.App.Helpers;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.ComponentModel;
 using System.Threading.Tasks;
 using Windows.System;
 
@@ -15,7 +12,7 @@ namespace Files.App.Actions
 {
 	internal class DeleteItemAction : ObservableObject, IAction
 	{
-		public IContentPageContext context = Ioc.Default.GetRequiredService<IContentPageContext>();
+		private readonly IContentPageContext context = Ioc.Default.GetRequiredService<IContentPageContext>();
 
 		public string Label { get; } = "Delete".GetLocalizedResource();
 
@@ -23,8 +20,7 @@ namespace Files.App.Actions
 
 		public HotKey HotKey = new(VirtualKey.Delete);
 
-		public bool IsExecutable => context.ShellPage is not null && context.SelectedItems is not null
-			&& context.SelectedItems.Any() && context.PageType is not ContentPageTypes.Home;
+		public bool IsExecutable => context.HasSelection;
 
 		public DeleteItemAction()
 		{
@@ -33,18 +29,14 @@ namespace Files.App.Actions
 
 		public async Task ExecuteAsync()
 		{
-			await RecycleBinHelpers.DeleteItem(context.ShellPage);
+			if (context.ShellPage is not null)
+				await RecycleBinHelpers.DeleteItem(context.ShellPage);
 		}
 
-		public void Context_PropertyChanged(object? sender, System.ComponentModel.PropertyChangedEventArgs e)
+		public void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
 		{
-			switch (e.PropertyName)
-			{
-				case nameof(IContentPageContext.SelectedItems):
-				case nameof(IContentPageContext.Folder):
-					OnPropertyChanged(nameof(IsExecutable));
-					break;
-			}
+			if (e.PropertyName is nameof(IContentPageContext.HasSelection))
+				OnPropertyChanged(nameof(IsExecutable));
 		}
 	}
 }
